requirements.txt
flask
opencv-python-headless
tensorflow-cpu
gunicorn
pillow
numpy
